+++ plugins/ping.js
@@ -13,6 +13,11 @@ let ping = {
         }
     }
 }
+
+// pong untuk yang interaktif
+
+const { performance } = require('perf_hooks');
+
 let pong = {
     name: 'pong',
     status: true,
@@ -22,9 +27,21 @@ let pong = {
         let message = update.message
         let text = message.content.text.text
 
-        if (this.regex.exec(text)) {  
-            tg.sendChatAction(message.chat_id)          
-            return tg.sendMessage(message.chat_id, '🐧 <b>Ping..uin!</b>', 'html', false, false, false, message.id)
+        if (this.regex.exec(text)) {
+            return tg.sendMessage(message.chat_id, 'Pingiiin...', false, false, false, false, message.id)
+                .then(result => {
+
+                    let t0 = performance.now();
+
+                    // console.log(result)
+                    tg.client.once('message' + result.id, (message) => {
+                        // console.log(message)
+                        let t1 = performance.now();
+                        let selisih = '<code>' + (t1 - t0).toLocaleString('id-ID') + "</code> milidetik."
+                        let pesan = '🐧 <b>Ping-uin!</b>\n👟 ... berlari dalam ' + selisih
+                        tg.editMessageText(message.chat_id, message.id, pesan, 'html')
+                    })
+                })
                 .catch(e => console.log(e))
         }
     }