+++ doc/history.md
@@ -27,6 +27,11 @@ Log history dari versi developtment
 + [.3.9] Load plugins dipengaruhi level debug
 + [.3.9] Config: input nomor hp dari sini
 
+`20 April 2021`
+
++ [.3.11] TGLib 1.3: handle diganti client, +getRemoteFile, +downloadFile
++ [.3.11] Plugins: .pong dengan responsif time
+
 ### v0.2 a
 
 `17 April 2021`